Combined prognosis and treatment decision nomogram (CPTDN)
(A) A CPTDN was built for predicting simultaneously disease-free survival (DFS) and treatment response. The calibration curves of the CPTDN were plotted for predicting the 3-year (B) and 5-year DFS rate (C) in three cohorts. The CPTDN could noninvasively recommend an optimal treatment regimen and predict patient prognosis. For example, consider a 40-year-old target patient with stage T3N1M0, a pre-EBV DNA of >4000 copy/mL, a Prognostic-score of 0.0, and a Predictive-score of -0.4; If the patient received ICT+CCRT, the patient’s total score would be 122 (21+15+41+45), for which the 5-year DFS rate is estimated at 81%. If the patient received CCRT, the patient’s score would be 147 (21+15+41+70), for which the 5-year DFS rate is estimated at 63%. The 5-year survival benefit of receiving ICT+CCRT vs. receiving CCRT is 18%, so the CPTDN would recommend ICT+CCRT and provide a 5-year DFS rate of 81%. CCRT, concurrent chemoradiotherapy; ICT, induction chemotherapy; DFS, disease-free survival; pre-EBV DNA, pre-treatment plasma Epstein-Barr virus DNA.